Improve memory view. Prevent it from drifting. Recenter it when the memory pointer gets too close to the edge.
Create a MemoryView class.
Don't set the dimensions of the memory SVG. Instead, set overflow to visible. This makes it easier to prevent it from drifting when the dimensions change.
Prevent the memory SVG from drifting when the dimensions change.
Hook up the reset view button.
When the memory pointer gets too close to the edge, autoscroll to it.
Show the memory pointer information in the state window.
Only update the memory view when something changes.