Convert the application to use a new App top-level React component.
Create a CodePanel component that creates the initial structure, which used to be in index.html. It doesn't do anything else. The GridView class still implements the contents of this panel.

Make React component classes/functions public. Remove helper functions that render them.

Add immer package to simplify state management in the App component.

In the App component, use class properties to avoid having to either make lambda expressions to call functions or bind this to each function.
Add babel-eslint plugin, to enable eslint to parse class properties.
Add the @babel/plugin-proposal-class-properties plugin, to enable babel to convert class properties to something more compatible.

eslint: change no-unused-vars to a warning.

GridView:
- Remove updateUndoButtonsCallback, which is no longer useful, since React allows us to avoid manually doing targeted updates to controls.
- Remove the timeoutID property. It never belonged here.

Remove the concept of isProgrammatic when setting source code. isProgrammatic was only ever true when initFinished was false. It was kind of redundant. Both things helped to prevent saving to localStorage immediately on loading the page, but neither are necessary anymore.

Note that this change is net-negative lines of code. App.jsx is almost exactly the same length that main.mjs was and code has been removed from other files.