Move undo/redo stacks from GridView to App component state.
When editing code in the Code panel, preserve whitespace and debug chars.
When resizing or resetting the code, preserve its minified or laid out status.
Improve the handling of code containing the debug character `.

When making the code bigger or smaller, offset the bottom half of the hexagon so that mirrors still work. The code still stays in the upper left corner. If it doesn't use wrapping, it would continue to work after resizing, as long as it wasn't truncated.

This makes the code cleaner and makes it easier to convert GridView to a React component.

The SourceCode class makes it easier to process the code and reduces the usages of Symbol.iterator from five to one.