Split up the Hexagony class into HexagonyState and HexagonyContext.
I split up the Hexagony class into things that evolve over time and would be stepped backwards if I implemented that, the HexagonyState and things that wouldn't be, the HexagonyContext. I'm thinking about how the step back feature would interact with the current edit-and-continue/undo/redo feature. I'm leaning towards them both being independent. For example, you pause execution, modify some code a few steps back from the current IP, then step backwards a few times and then forwards. Stepping backwards would not undo the code change you made. When you step forwards again over the changed code, it would do something different. Code changes can be undone/redone as normal, while the program is running, with the current restriction that any undo/redo action which changes the hexagon size is not allowed. I'm planning to keep that restriction.

Moved executionHistory from App to HexagonyState, since changes to it will need to be undone to implement step back.

HexagonyState and Memory are immerable, which allows them to be used either as normal objects or as immutable objects with copy-on-write.