Remove some old irrelevant code that was triggering the memory view to recenter every time memory data changed.
The memory panel used to stick around between programs, even though you couldn't see it, and when the memory object reference was different, it would recenter its view, because a new program had started. With the new immutable objects, this was triggering a recentering every time memory was written to.