For dark mode, make the foreground color a bit darker than pure white, except in the hexagons.
Previously, it was noticeably too bright on the about page.

In the hexagons, maximum contrast is required, particularly for reading cell text on top of the active execution cell background color.
Also:
Remove --color
Rename --hover to --header-hover-inactive
Rename --hover-active to --header-hover-active
Add header-disabled-color.
Move header and direction picker collors to :root. This will be useful for the about page, which describes toolbar icons.