Improve support for keyboard navigation with tab.
The user can now tab in and out of the code panel.
The user can now pan the memory panel, when it has keyboard focus, with the arrow keys and zoom it with numpad + and -.
Move the memory view before the memory reset button, so that it's first in the tab order.