Update CodePanel pan zoom to fix zooming issues.
Placing the pan zoom on the SVG allows zooming to work correctly. Otherwise, it gets confused due to the styling that I use to keep the contents of the code panel centered when the window resizes. Use the code panel as the keyboard owner of the pan zoom, a concept I added, for better keyboard navigation. The keyboard focus rectangle will be more clear and the mouse wheel will work throught the code panel.