Allow horizontal scrolling to see the content in the code panel, if necessary.
The essential area for the code panel includes closest two rows on each of the adjacent hexagons. If this area doesn't fit either horizontally or vertically, then users will be able to scroll to see the rest of it, but no more on that dimension. Users will only see past the area horizontally, if their screen is big enough. Users will only see past the area vertically if their display is wide enough to use the two column layout and the hexagon is small relative to the two rows on the right that share the same grid layout slot. This will help to avoid getting disoriented when scrolling around a large hexagon. The vertical scrolling behavior has been implemented this way for a long time.

Updated GridView.ts to determine the essential area.

Removed code that used the panzoom library for horizontal scrolling in the code panel. This solution feels much better and the browser does all of the work, which helps it to feel more natural.